Allow to specify output locale for the inline report both via command line and webapp.
Pass the locale through all the way to the InlineReport object and use it when formatting fact values for output.

Also switch the inline transform used depending on the decimal_symbol of that locale (expected to be either comma or full-stop and we raise an Exception if it is not).