No need to log a warning if we're dropping a section (table) with no facts ... that is normal.
When merging table sections together with non-table sections, make sure to only include table sections that exist. Previously, if a table had no data, the old report non-tabular section was included, leading to broken HTML (a table with no rows), as a table was expected but none found.

Examples of the errors:

[lxml.SCHEMAV_ELEMENT_CONTENT] XML file syntax error Element '{http://www.w3.org/1999/xhtml}thead': Missing child element(s). Expected is ( {http://www.w3.org/1999/xhtml}tr )., line 3339, path 'body/div[30]/table/thead', xpath '/*/*[2]/*[30]/*[2]/*[1]'

[lxml.SCHEMAV_ELEMENT_CONTENT] XML file syntax error Element '{http://www.w3.org/1999/xhtml}tbody': Missing child element(s). Expected is ( {http://www.w3.org/1999/xhtml}tr )., line 3342, path 'body/div[30]/table/tbody', xpath '/*/*[2]/*[30]/*[2]/*[2]'